backend/requirements.txt
# GAIM Lab v3.0 Backend Dependencies

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Async & Background Tasks
celery[redis]>=5.3.0
redis>=5.0.0

# AI & ML
google-generativeai>=0.3.0
faster-whisper>=1.0.0
openai-whisper>=20231117
mediapipe>=0.10.9
deepface>=0.0.80

# Audio/Video Processing
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.11.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data & Utils
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# PDF Generation
playwright>=1.40.0
weasyprint>=60.0

# GraphQL (optional)
strawberry-graphql>=0.220.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
